+++ Project/EnemyComponents/IEnemyState.cs
@@ -19,7 +19,7 @@ namespace Project1.EnemyComponents
         int Size { get; set; }
         void Draw(SpriteBatch spriteBatch, Vector2 position);
         void Update();
-        void TakeDamage(int damage);
+        void TakeDamage(double damage);
        
     }
 }
